REF: test_to_timestamp
